+++ packages/components/src/utils/form/isRequiredAndParentNotRequired.ts
@@ -1,7 +1,7 @@
-import { isParentFieldsetWrapperRequired } from './isParentFieldsetWrapperRequired';
+import { isParentFieldsetRequired } from './isParentFieldsetRequired';
 import type { HTMLElementWithRequiredProp } from './isRequired';
 import { isRequired } from './isRequired';
 
 export const isRequiredAndParentNotRequired = (element: HTMLElement, child: HTMLElementWithRequiredProp): boolean => {
-  return isRequired(child) && !isParentFieldsetWrapperRequired(element);
+  return isRequired(child) && !isParentFieldsetRequired(element);
 };